Moving ias-slot-setup back out and setting them up with valid sizes
